chore: HomeViewController 병합
