+++ ios/Daangn/Daangn/Controller/HomeViewController.swift
@@ -8,18 +8,47 @@
 import UIKit
 
 final class HomeViewController: UIViewController {
-    private let category = CategoryViewController()
-    private let configureTown = ConfigureTown()
+    private let collectionView = ProductListCollectionView()
+    private lazy var dataSource: ProductListDataSource? = ProductListDataSource(collectionView)
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setNavigationBar()
+        setLayout()
+        applyUpdatedSnapshot()
+    }
+    
+    private func setLayout() {
+        view.addSubview(collectionView)
+        NSLayoutConstraint.activate([
+            collectionView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
+            collectionView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
+            collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            collectionView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
+        ])
+    }
+    
+    private func applyUpdatedSnapshot() {
+        var snapshot = NSDiffableDataSourceSnapshot<ProductListSection, ProductListItem>()
+        snapshot.appendSections([.product, .load])
+        // TODO: 임시 코드 수정
+        let products = (1...100).map { ProductListItem.product($0) }
+        snapshot.appendItems(products, toSection: .product)
+        if true { snapshot.appendItems([.load], toSection: .load) }
+        dataSource?.apply(snapshot, animatingDifferences: true)
+    }
+    
     private let town = "역삼1동"
-    func menuHandler(action: UIAction) {
+    
+    private func menuHandler(action: UIAction) {
         Swift.debugPrint("Menu handler: \(action.title)")
     }
     
-    func moveToTownPicker(action: UIAction) {
-        present(configureTown, animated: true)
+    private func moveToTownPicker(action: UIAction) {
+        present(ConfigureTown(), animated: true)
     }
     
-    func navBarItem() {
+    private func setNavigationBar() {
         let leftButton = UIBarButtonItem(title: town, style: .plain, target: self, action: nil)
         let rightButtonImage = UIImage(systemName: "line.3.horizontal")
         let rightButton = UIBarButtonItem(image: rightButtonImage, style: .plain, target: self, action: #selector(moveToCategory))
@@ -31,19 +60,15 @@ final class HomeViewController: UIViewController {
         configureNavBarButtonColor(left: leftButton, right: rightButton)
     }
     
-    func configureNavBarButtonColor(left: UIBarButtonItem, right: UIBarButtonItem) {
+    private func configureNavBarButtonColor(left: UIBarButtonItem, right: UIBarButtonItem) {
         navigationItem.rightBarButtonItem = right
         navigationItem.leftBarButtonItem = left
         navigationItem.rightBarButtonItem?.tintColor = .black
         navigationItem.leftBarButtonItem?.tintColor = .black
     }
     
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navBarItem()
-    }
-    
     @objc func moveToCategory() {
-        self.navigationController?.pushViewController(category, animated: false)
+        let nextViewController = CategoryViewController()
+        self.navigationController?.pushViewController(nextViewController, animated: true)
     }
 }
